+++ src/delivery/client.ts
@@ -1,42 +1,58 @@
 import { DeliveryClientConfig, useDeliveryConfigDefaults } from "../config/config";
 import { DeliveryServiceClient } from "@fraym/crud-proto";
 import { credentials } from "@grpc/grpc-js";
-import { createCrudData, CreatedCrudData } from "./create";
-import { updateCrudData } from "./update";
+import { createCrudData, CreateResponse } from "./create";
+import { updateCrudData, UpdateResponse } from "./update";
 import { deleteCrudData } from "./delete";
 import { getCrudData } from "./getData";
 import { GetCrudDataList, getCrudDataList } from "./getDataList";
 import { Filter } from "./filter";
 import { Order } from "./order";
+import { AuthData } from "./auth";
+import { EventMetadata } from "./eventMetadata";
 
 export interface DeliveryClient {
-    create: (
-        tenantId: string,
-        type: string,
-        data: Record<string, any>,
-        id?: string
-    ) => Promise<CreatedCrudData>;
-    update: (
-        tenantId: string,
-        type: string,
-        id: string,
-        data: Record<string, any>
-    ) => Promise<void>;
-    delete: (tenantId: string, type: string, id?: string, filter?: Filter) => Promise<void>;
     getData: <T extends {}>(
-        tenantId: string,
         type: string,
+        authData: AuthData,
         id: string,
+        filter?: Filter,
         returnEmptyDataIfNotFound?: boolean
-    ) => Promise<T>;
+    ) => Promise<T | null>;
     getDataList: <T extends {}>(
-        tenantId: string,
         type: string,
+        authData: AuthData,
         limit?: number,
         page?: number,
         filter?: Filter,
         order?: Order[]
     ) => Promise<GetCrudDataList<T>>;
+    create: <T extends {}>(
+        type: string,
+        authData: AuthData,
+        data: Record<string, any>,
+        id?: string,
+        eventMetadata?: EventMetadata
+    ) => Promise<CreateResponse<T>>;
+    update: <T extends {}>(
+        type: string,
+        authData: AuthData,
+        id: string,
+        data: Record<string, any>,
+        eventMetadata?: EventMetadata
+    ) => Promise<UpdateResponse<T>>;
+    deleteDataById: (
+        type: string,
+        authData: AuthData,
+        id: string,
+        eventMetadata?: EventMetadata
+    ) => Promise<number>;
+    deleteDataByFilter: (
+        type: string,
+        authData: AuthData,
+        filter?: Filter,
+        eventMetadata?: EventMetadata
+    ) => Promise<number>;
     close: () => Promise<void>;
 }
 
@@ -52,51 +68,77 @@ export const newDeliveryClient = async (config?: DeliveryClientConfig): Promise<
         }
     );
 
-    const create = async (
-        tenantId: string,
-        type: string,
-        data: Record<string, any>,
-        id: string = ""
-    ) => {
-        return await createCrudData(tenantId, type, data, id, serviceClient);
-    };
-
-    const update = async (
-        tenantId: string,
-        type: string,
-        id: string,
-        data: Record<string, any>
-    ) => {
-        return await updateCrudData(tenantId, type, id, data, serviceClient);
-    };
-
-    const deleter = async (
-        tenantId: string,
-        type: string,
-        id: string = "",
-        filter: Filter = { fields: {}, and: [], or: [] }
-    ) => {
-        return await deleteCrudData(tenantId, type, id, filter, serviceClient);
-    };
-
     const getData = async <T extends {}>(
-        tenantId: string,
         type: string,
+        authData: AuthData,
         id: string,
+        filter: Filter = { fields: {}, and: [], or: [] },
         returnEmptyDataIfNotFound: boolean = false
-    ): Promise<T> => {
-        return await getCrudData<T>(tenantId, type, id, returnEmptyDataIfNotFound, serviceClient);
+    ): Promise<T | null> => {
+        return await getCrudData<T>(
+            type,
+            authData,
+            id,
+            filter,
+            returnEmptyDataIfNotFound,
+            serviceClient
+        );
     };
 
     const getDataList = async <T extends {}>(
-        tenantId: string,
         type: string,
+        authData: AuthData,
         limit: number = 0,
         page: number = 1,
         filter: Filter = { fields: {}, and: [], or: [] },
         order: Order[] = []
     ): Promise<GetCrudDataList<T>> => {
-        return await getCrudDataList<T>(tenantId, type, limit, page, filter, order, serviceClient);
+        return await getCrudDataList<T>(type, authData, limit, page, filter, order, serviceClient);
+    };
+
+    const create = async <T extends {}>(
+        type: string,
+        authData: AuthData,
+        data: Record<string, any>,
+        id: string = "",
+        eventMetadata: EventMetadata = { causationId: "", correlationId: "" }
+    ) => {
+        return await createCrudData<T>(type, authData, data, id, eventMetadata, serviceClient);
+    };
+
+    const update = async <T extends {}>(
+        type: string,
+        authData: AuthData,
+        id: string,
+        data: Record<string, any>,
+        eventMetadata: EventMetadata = { causationId: "", correlationId: "" }
+    ) => {
+        return await updateCrudData<T>(type, authData, id, data, eventMetadata, serviceClient);
+    };
+
+    const deleteDataById = async (
+        type: string,
+        authData: AuthData,
+        id: string,
+        eventMetadata: EventMetadata = { causationId: "", correlationId: "" }
+    ) => {
+        return await deleteCrudData(
+            type,
+            authData,
+            id,
+            { fields: {}, and: [], or: [] },
+            eventMetadata,
+            serviceClient
+        );
+    };
+
+    const deleteDataByFilter = async (
+        type: string,
+        authData: AuthData,
+        filter: Filter = { fields: {}, and: [], or: [] },
+        eventMetadata: EventMetadata = { causationId: "", correlationId: "" }
+    ) => {
+        return await deleteCrudData(type, authData, "", filter, eventMetadata, serviceClient);
     };
 
     const close = async () => {
@@ -104,11 +146,12 @@ export const newDeliveryClient = async (config?: DeliveryClientConfig): Promise<
     };
 
     return {
-        create,
-        update,
-        delete: deleter,
         getData,
         getDataList,
+        create,
+        update,
+        deleteDataById,
+        deleteDataByFilter,
         close,
     };
 };