+++ src/delivery/create.ts
@@ -1,29 +1,52 @@
 import { DeliveryServiceClient } from "@fraym/crud-proto";
+import { AuthData, getProtobufAuthData } from "./auth";
+import { EventMetadata } from "./eventMetadata";
 
-export interface CreatedCrudData {
-    id: string;
+export type CreateResponse<T extends {}> = CreateSuccessResponse<T> | CreateValidationResponse;
+
+export interface CreateSuccessResponse<T extends {}> {
+    data: T;
 }
 
-export const createCrudData = async (
-    tenantId: string,
+export interface CreateValidationResponse {
+    validationErrors: string[];
+    fieldValidationErrors: Record<string, string>;
+}
+
+export const isCreateSuccessResponse = <T extends {}>(
+    response: CreateResponse<T>
+): response is CreateSuccessResponse<T> => {
+    return response.hasOwnProperty("data");
+};
+
+export const isCreateValidationResponse = <T extends {}>(
+    response: CreateResponse<T>
+): response is CreateValidationResponse => {
+    return !response.hasOwnProperty("data");
+};
+
+export const createCrudData = async <T extends {}>(
     type: string,
+    authData: AuthData,
     data: Record<string, any>,
     id: string,
+    eventMetadata: EventMetadata,
     serviceClient: DeliveryServiceClient
-): Promise<CreatedCrudData> => {
+): Promise<CreateResponse<T>> => {
     const requestData: Record<string, string> = {};
 
     for (const key in data) {
         requestData[key] = JSON.stringify(data[key]);
     }
 
-    return new Promise<CreatedCrudData>((resolve, reject) => {
+    return new Promise<CreateResponse<T>>((resolve, reject) => {
         serviceClient.createEntry(
             {
-                tenantId,
                 type,
+                auth: getProtobufAuthData(authData),
                 data: requestData,
                 id,
+                eventMetadata,
             },
             (error, response) => {
                 if (error) {
@@ -31,8 +54,22 @@ export const createCrudData = async (
                     return;
                 }
 
+                if (response.validationErrors || response.fieldValidationErrors) {
+                    resolve({
+                        validationErrors: response.validationErrors,
+                        fieldValidationErrors: response.fieldValidationErrors,
+                    });
+                    return;
+                }
+
+                const data: any = {};
+
+                for (const key in response.newData) {
+                    data[key] = JSON.parse(response.newData[key]);
+                }
+
                 resolve({
-                    id: response.id,
+                    data,
                 });
             }
         );